Don't check for line presence in the latest commit
This fixes an issue that Hound wound't comment when multiple commits got pushed
to a PR at the same time. As it was, Hound would check if only the head commit
included the line with violation -- resulting in no comment if the violation
came in the previous commit.

Instead, simply check if the violation on the line was previously unreported.