Use default show_exceptions value of false
Enabling this hides exceptions during integration tests behind an HTML
500 page, which makes debugging much harder. It can also hide errors, as
the test can render a 500 and still pass.

I didn't have redis running locally, which caused one test to fail. With
the previous setting, it took me 15 minutes to find the failure. With
the new setting, it immediately prints an error about redis.