Rename "activity" report to "north star" report
Use "North Star" term to match the thoughtbot growth team's terminology
for [the only metric that really matters][only].

[only]: https://medium.com/@joshelman/the-only-metric-that-matters-ab24a585b5ea

> How many people are really using your product?

Identify the interaction that represents the moment users really get value from
your product and identify the threshold for how often that interaction is
appropriate within the context of the product.

For example, Hound's "north star metric" is:

> Repos with builds per week

When Hound reviews a pull request or a new commit to a pull request, it
"builds." This is the "aha moment" when a user is getting value from the
product. We care about this number because most teams relax on the weekends.