Allow a user to exclude certain JavaScript files
Allows a user to configure hound to ignore specified JavaScript files.
This follows the "jshintignore" convention. If a .jshintignore file is
found and the user has not overwritten the path in the hound
configuration, it will automatically be used.

Otherwise, a user can specify the path to a file which specifies which
files to exlude during linting.

Example configuration:

```yaml
java_script:
  enabled: true
  ignore_file: ".javascript_ignore"
```

.javascript_ignore
```
app/assets/javascripts/**.js
``