Rescue requests that set status
Setting a pull request status requires at least "push" permissions. When we
create "Services" team for Hound we only set "pull" permissions, which is not
enough to create statuses.

Note: that this affects about half the organizations, since we will add Hound to
an existing team, if it meets certian requirements.