Account for comments that change patch position
When fetching comments on a pull request, the GitHub api includes two
values indicating the position of the comment in the patch: `position`
and `original_position`. Most of the time, `position` is nil, and
`original_position` contains the relevant value. However, if an event
causes the patch to change, e.g. a new commit is pushed, both `position`
and `original_position` will contain values. This change looks for the
presence of `position`, using it for the comparison when appropriate.