Fix SCSS linter creating erroneous violations
The SCSS linter stores violations in an instance variable[1], and
doesn't clear them between runs or provide a public interface for
clearing them. This led to erroneous violations being reported on files
when a violation in another file happened to be on the same line number
that's included in the diff.

This change uses a newly created linter for every file that's checked,
in order to avoid the responsibility of clearing the state between each
run.

Will close #562

[1] https://github.com/causes/scss-lint/blob/8319483e3686e1956613069e6a02a56e62db74bc/lib/scss_lint/runner.rb#L12