Read MAX_COMMENTS using ENV#fetch
Given a PR with violations in the diff, if `MAX_COMMENTS` is not set Hound will
not report any violations. This is a kind of silent failure, which I think we
should avoid.

Also put the value of `MAX_COMMENTS` into a constant evaluated at boot time in
order to fail fast if it is not set.