Return unique list of repos from repos.json
Currently, it is possible for there to be duplicate `memberships` tying a `user`
to `repos` and as a result, the user can sometimes see duplicate `repos` on
houndci.com. This update solves the UI problem. We'll have to dig a bit deeper
to figure out where the duplicates are coming from in the first place.

Part of
https://trello.com/c/GnvVhljg/532-repos-json-endpoint-returns-duplicate-repos