Only spin spinners when they are visible
Turns out, the CPU load was caused by the many animated spinners that
are hidden on the page at any given time. Every repo has a spinner
that's only shown when it's being activated or deactivated.

With `visibility: hidden`, the browser was still animating these
spinners even when they weren't visible. You can see that in the Chrome
devtools by selecting a `.repo::before` pseudo-element, and you'll see a
spinning outline of where the component is hidden.

The fix was to only activate spinning on processing elements, instead of
all elements that may or may not be visible.