Automatically retry failed jobs
This changeset removes resque-retry and resque-timeout,
which are incompatible with resque under ActiveJob.
They are replaced with a new ApplicationJob class
and a rewrite of the Retryable module.

A new scheduler task is added to the Procfile,
which runs an instance of resque-scheduler and
executes delayed jobs.

Exceptions in jobs are no longer explicitly
caught and sent via `Raven.capture_exception`,
and instead fall through to the default
Resque::Failure::Sentry handler.