Return an existing subscription during activation
We've recently encountered a situation where we believe the customer had
a subscription, but was attempting to activate again -- this can also be
caused by duplicate, simultaneous requests for the same user.

There is not reason to not return the existing subscription, if it
exists and avoid charging the user again via Stripe.