Handle single (non-array) SCSS excludes
Users can provide excludes for SCSS linters as an array:

    exclude:
      - lib/**
      - spec/assets/**

or as a string:

    exclude: spec/assets/**

We need to run the config hash through SCSSLint's methods to normalize
all of the options. In order to do that, we need to write to a file and
provide a path to that file.

Related error: https://app.getsentry.com/hound-1/production/group/63380345/