Remove expired token when running a build
If we encounter an expired token,
then the job should retry with another token.
When there are no more users with tokens,
the jobs should run with Hound's token,
and if that fails due to permissions
then the job retries should stop.

Due to token encryption, to clear out a token,
and set it to `NULL` in the DB,
it is necessary to set it directly,
and avoid hitting `User#token=`.