Remove unreachable repo from user's repos
When a repo cannot be reached with a user's token,
causing a 404 Not Found error,
that means that user has been barred access to it.

By removing the repo from user's associated repos,
the job will be retried with a different user's token,
or fallback to Hound's.

This should fix numerous 404 errors that are causing jobs to fail.