Require Bundler >= 1.10
As of Bundler 1.10, Bundler now includes an additional section in
Gemfile.lock, "BUNDLED WITH", that tells it which version was used to
run `bundle install`:

<bundler/bundler@a264e6b>

Unfortunately, on versions prior to 1.10, this section is missing, so
Gemfile.lock will change every time you run `bundle` until you upgrade
to Bundler 1.10.

For this reason, bump the dependency on Bundler to force people to
upgrade.