Switch from Travis to Circle for CI testing
- Copy .sample.env to .env before testing to ensure environment
  variables are set correctly
- Remove travis configuration file
- Switch out README status badges

- Remove konacha runner in favor of Jasmine

CircleCI was auto-detecting Konacha in the app and attempting to run it
as part of the build, with `rake konacha:run --trace`.

Since Konacha was not set up with the correct dependencies,
the rake task was failing.

Meanwhile, we've been using `rake spec:javascript` to run the javascript
specs from our Rakefile. Removing Konacha makes sure we're using
consistent testing commands between CI and local machines.