Extract PR status setting to CommitStatus class
If a build was processed with a user's token,
then it should set the final PR status using that token.
We now store the user whose token was used,
or default to @houndci's.

This solves the issue where the app was always setting
the final success status with `houndci`'s token,
instead of user's token (if available).

This also gets status setting code out of `BuildRunner`
and into its own `CommitStatus` class.