Include .env in the repo
Moves .sample.env to .env. This file has all the config necessary to get
running in development and `cp .sample.env .env` was in `bin/setup`. If
this file just exists as `.env` in the repo, then the setup step is not
needed and things should Just Work™ out of the box. The dotenv [docs][1]
even recommend doing this to ease new developers onto the project.

Additionally, our heavy use of `ENV.fetch` is now causing the test suite
to [fail][2] on CircleCI, because the database is being setup before the
`pre` hook fires. With `.env` the test suite will work fine without this
setup.

Local customizations can be added to `.env.local`, which is ignored.

[1]: https://github.com/bkeepers/dotenv#should-i-commit-my-env-file
[2]: https://circleci.com/gh/thoughtbot/hound/202