Add an admin dashboard using Administrate
The admin dashboard displays and allows the user to manage
bulk subscriptions.

Why?

Until now, bulk subscriptions had to be managed through the console.
Adding a dashboard makes it easier to add, remove and edit bulk
subscriptions.

This admin dashboard also gives us a chance to try out Administrate in
production before releasing it to the world.

We aren't adding other dashboards at the moment, because:

- Large size of most tables make Administrate prohibitive to use for
  most tables, until filtering is implemented
- Bulk Customers is a small table, and managing bulk customers is one of
  the common admin tasks

Specific Changes:

- Add Administrate to gemfile
- Run `rails g administrate:install`
- Remove unused controllers and dashboards
- Remove generated comments
- Add authorization logic for admin dashboards
  Admins can be added or removed by setting an environment variable,
  `ADMIN_GITHUB_USERNAMES`, as a csv list
- Add `BulkCustomer#to_s`, so bulk customers are displayed nicely in
  the dashboard
- Customize routing to get rid of empty `Admin::ApplicationController`
  subclasses
- Remove limit on index page record count