Don't fail on violations
This was enabled on Hound itself in order to vet the feature before
releasing it to the world. The Hound team treats violations as
suggestions and sometimes decides to ignore them. In these cases,
failing the PR does not make sense. The feature has been working well
for quite awhile now, so it can be disabled here.