Merge RuboCop overrides first
There is some important code that is run in
`RuboCop::ConfigLoader.merge_with_defaults` -- it checks for presence of
`DisabledByDefault` and disables all of the rules but the overrides.

Because we were calling `RuboCop::ConfigLoader.configuration_from_file`
first (and it calls `merge_with_defaults` in turn) and then merging the
custom config, we were not preserving the `DisabledByDefault` logic.

Fixes #1049