Report invalid HAML format as violation
HamlLint includes the line number in the exception when it encounters an
invalid Haml format.

Normally, HamlLint's lines start at 1. However, this exception is raised
by the Haml gem when parsing the file, and HamlLint wraps it.
Haml reports the first line as 0 (zero based line numbering).
Thus, we need to add 1 to the line number when reporting this exception
as violation.