Don't pass filename to HamlLint
When we pass a filename to HamlLint and RuboCop linter is enabled then
HamlLint will try to create a Tempfile in the directory that it gets
from the file's path. It won't exist on Heroku, so it throws an
exception.

HamlLint still works without a filename, so let's not pass it in.

Fixes the following issue we've been seeing:
```
No such file or directory @ rb_sysopen - app/views/articles/_form.html.haml.haml_lint.tmp20160205-6932-alnw3e
```