Add script to show what isn't on production
The goal to get a bird's eye view of what needs to be deployed to
production.

The script assumes all of the hound related repositories are in the same
directory as the main one (this one). It then loops them and attempts to
compare production/master to origin/master, showing the results.