Don't rescue Octokit::NotFound when setting status
Introduced in e9836eea64baefb0699952c5f35fd65a9ca2df41, the logic was in
regard to adding Hound user to teams. Since, we are not doing that
anymore this is not longer relevant to the original problem.

Currently, there is an issue where we choose a user token to handle the
repo, since a repo could be public, the user's token will work for
processing the build/review.
But, it will fail silently when it tries to set the status. This is the
cause of the GitHub status sometimes not being set, a token that was
picked does not have status setting privileges.
If we raise, the pending status will fail and the job will retry using a
new token.