Support .eslint.js
ESLint is a bit complicated, and one of its complications, is how it handles config files.

In hound currently, we pass the config file to the queue and assume it is JSON.

This change, simply passes through file content as well as the file name so that when eslint goes to run it will know what to do.

This enables support for .eslint.js and .eslint.yaml, as well as .eslint.json and .eslint

In the future, we should probably find a better way to pass around configs? but this should work for now.