Ensure uniqueness of repos by github_id
Previously, we were ensuring repo uniqueness by name, but since we are
tightly coupled to GitHub, we should be using their ids.

There was a recent bug that led to this:

* On GitHub, `thoughtbot/constable-api` repo was renamed to `thoughtbot/constable`
* On GitHub, `thoughtbot/constable` repo was renamed to `thoughtbot/constable-react`
* When we were syncing repos, we would find the `thoughtbot/constable`
  and create a membership for it
* Then we would see the `thoughtbot/constable-react`. Nnot find it
  by name, then look it up by `github_id` and pull up the same
  `thoughtbot/constable` record and try to create a membership,
  which would fail.

If we relied on `github_id` for uniqueness, we would eventually update
the names to match what GitHub has.