Add file_reviews.linter_name (#1102)
[Related Github issue](https://github.com/houndci/hound/issues/1101)

Generally speaking, for any given language, Hound supports only one linter. For
example, Rubocop will handle all `.rb` files; SCSS-lint will lint `.scss` files
and so on.

And then there's JavaScript.

For JavaScript, Hound supports three linters: JShint, JSCS, and ESlint,
each of which can be individually enabled or disabled. If both JShint and
JSCS are enabled, however, Hound code reviews will never complete for a given
build, and this is because `CompleteFileReivew#file_review` will only look up
a `FileReview` by its `filename` attribute, and when multiple linters are
working on a single file, that is not enough information. The second through nth
`FileReivew` for that file never gets found, and its `completed_at` never gets
updated. When that happens, `Build#complete?` always returns false.

This change creates a migration that adds a new column to the `file_reviews`
table, which will store the name of the linter assigned to a given file review.
The corresponding `FileReview` model will acquire a model-level validation to
ensure the presence of this new linter-name attribute.

It then updates `Linter::Base#enqueue_job` to add the linter_name attribute to
the hash of review-job attributes so that linters can pick up this new
attribute. When their reviews are complete, the linters can be configured to
pass the linter name back to Hound, along with any errors and the filename, just
as before.

The last part of this solution involves updating
`CompleteFileReview#file_review` to anticipate an optional `linter_name`
attribute in its dequeued job attributes, which it will use to perform a more
precise search of `FileReview` objects in the relevant build. In order to avoid
a large coordinated upgrade effort, this method will fall back to the previous
locator logic, which only operates on the job's `filename` attribute.