Add index for builds pr number and sha
When completing a file review job, we find the associated build record
by pull request number and commit sha. With about 1.6 million records
in the table, this query is averaging 1820ms to execute.

This adds an index on the pull_request_number/commit_sha pair to improve
the performance of that query.