HoundConfig merges default and user configurations
Previously, `hound_config#enabled_for?` would return `true` for any
value that was not:

 - Explicitly set to false in the supplied config file
 - A language marked as beta

Additionally, HoundConfig didn't properly account for aliases. This
produced several bugs (all of which are now resolved) including:

 - Users disabling aliases (e.g. javascript) but still running the
   linter(e.g. jshint)
 - Users specifying ignore files that don't get properly used
 - .hound.yml files parsed as Rubocop config files if they only include
   alias or beta linter keys

Instead of trying to check for each linter name every time we get
something out of the config, we now start with a default configuration
and merges the user-specified configuration into it using a series of
transforms.

Side effects:

 - Remove linter_names from linter classes