enabled_for? returns false by default
Previously, `hound_config#enabled_for?` would return `true` for any
value that was not:

 - Explicitly set to false in the supplied config file
 - A language marked as beta

This changes `enabled_for?` to instead check if it is a supported linter
that is not in beta. This also fixes checks for legacy config files
which would have returned `true` if the config file only had keys for
beta linters.