Calculate config names from their class names
Previously you would create a `Ruby` config by doing:

    Config::Ruby.new(hound_config, "ruby")

This seems unnecessary since the information that it is a ruby linter
config is right in the class name. It also implies that you might want
to do something like:

    Config::Ruby.new(hound_config, "python")

But this just changes the way the object calculates `linter_names`, it
doesn't actually provide additional functionality that the app is using.

Besides making the API for creating configs simpler, this change allows
us to more easily fetch aliases from a type of config, since
`linter_names` can become a class method.