Add disabled_for? to HoundConfig
Previously, HoundConfig only had one way of checking the status of a
config option (`enabled_for?`), but it didn't account for aliases. This
could result in someone disabled an alias (like `javascript` but still
running the actual linter `jshint`).

This adds an additional method `disabled_for?` and sets both of these
methods to **only** check their respective states (in other words,
`enabled_for?` doesn't care if something is disabled, and
`disabled_for?` doesn't care if something is enabled. This allows
`CheckEnabledLinter` to first check whether a particular config is
disabled before it checks whether the config is explicity
enabled or default.