Add BuildOwnerHoundConfig service
Owners can now have a config repository configured. When given an
`owner` `BuildOwnerHoundConfig` will fetch the HEAD commit of the repo
and give back a `HoundConfig` built from it. If the owner does not have
a config repo or it is disabled `BuildOwnerHoundConfig` will return
nil.

We also set any owner with an active repo to use owner level
configuration pointed at `houndci/legacy-config` in order to keep
anyone's builds from breaking when we eventually remove the hound
defaults from this app.