Check for conflicting linters + resolve conflicts
We now have multiple linters which can lint the same file (e.g. `jshint`
and `eslint`). It probably only makes sense to use one of these at a
time, so Hound now checks to see if there are any conflicts between
enabled linters (specifically if eslint is enabled and javascript is not
explicitly disabled) and disables the conflicts explicitly.

Other changes:

 - Turn on `jshint` by default again (because I believe this is the
   expected behavior)