Silence 404 errors when setting status
Sometimes users will change `houndci-bot`'s permissions so that it
cannot set the status of the GitHub PR. This generates a lot of
failure noise in our job queues.

This actually doesn't affect the functionality of Hound that much,
since it will still review the PR and make comments when violations are
found. The PR status just won't be set/updated.