Simplify enabled linter checking
Removes `CheckEnabledLinter` in favor of performing the check inline.
`CheckEnabledLinter` was capable of supporting many config objects, each
with their own `HoundConfig`. Over time the actual usage of it has been
reduced to checking a single linter name given a single `HoundConfig`,
which simplifies the process so much that having a dedicated class is
overkill.