Speed Improvements for CommitFile and Patch models. Closes #1223
`CommitFile.line_at` has a point of improvements changing the way
we create and access the data structure, from an array to hash, this
means an improvement from O(n) to O(1).

In order to get this improvement we need to change how
`Patch.changed_lines` construct their data structure.