Add lib to eagerload paths
This fixes lib files not being available in Rails 5 production without
explicitly requiring them.

This comes with the caveat that everything in lib will be loaded when the
application boots, but in our case that's just a few rake tasks so we aren't too
concerned about it for the convenience this provides.