Don't fetch config when there is no owner
There was a infinite-loop situation, where we would attempt to keep
calling `content` (and subsequently `owner_config`) for each owner.
Since we moved to `MissingOwner`, we'd pass that in when there was no
owner, and thus loop infinitely passing around `MissingOwner` instance.

This prevents the above situation and returns empty config for
`MissingOwner`.