Dry up testing common linter behavior
While reviewing the specs for the suite of linters there is common
behavior that is repeatedly tested that is ripe for extraction to
something like a shared example.

To reduce boilerplate I've started with one quick win - a small test to
assert that linters allow only certain types of files and/or extensions.