Masquerading as a user
This adds a route that admins can go to that will put them in a mode where they
masquerade as the user they provided. They can then browse hound as if they were
that user, and then click "Stop Masquerading" to return to their own account.

Also add an administrate sidebar partial since we now have more than just
administrate in the admin namespace, we need to exclude it from the sidebar
manually since administrate does not provide a way to exclude it.

There's some work to solve this but hasn't been merged yet.
https://github.com/thoughtbot/administrate/pull/514

https://trello.com/c/h9VtPK6F/800-as-an-admin-i-would-like-to-masquerade-as-a-user-so-that-i-may-help-debug-issues-that-they-are-running-into