Rescue & handle RecordNotUnique error in repo sync
The `RepoSynchronization` service object has recently run into an issue
where a user has had a repo owner with a different github_id, but a name
for the Owner/Org that is the same as a previous Owner record. This has
caused an exception during the synchronization process that was not
recovered.

This commit will rescue that exception in the RepoSynchronization service
object, and will continue the `each` loop if it encounters it.

Down in the Owner model class, however, the upsert that creates this
exception will send the error to Raven along with some additional info
on the offending repo - the github_id and the name.