Sanitize Flake8 config
The configuration that Flake8 accepts from the config file can have
certain options on multiple lines, like:
```
[flake8]
ignore =
  # some comment
  E221,
  # another comment
  E222
```

The library we use to parse the ini file isn't able to handle that case
and errors out.

This solution puts the multi-line options on one line,
as well as removes all the comments from the config file.