Performance tune the repos endpoint
Fixes a handful of N+1 issues for the repos.json endpoint.

Also removes the `price_in_dollars` fields from the response. It's
calculated from the old $12/repo business logic. Eventually the whole
`Plan` object should be removed, but that's out of scope for this PR.

For my user, this reduces the number of queries from 1391 to 7.