Mark images as presentational
The purpose of these images are decoration because the content meant for
the user is already available through each `h5` sibling. By providing an
empty `alt` attribute on `img` elements, we mark them as presentational
to assistive technology. This means we won't force repetitive, duplicate
content onto the person consuming the page.