Pull normalize.css through Rails Assets
The normalize-rails gem has fallen out of date. The current version of
normalize.css is 7.0.0, but the latest version of normalize-rails is
4.1.1.

There have been repeated requests by collaborators of normalize-rails to
be given permissions to deploy new versions, but have gone unfulfilled.

With around 7 months of waiting, it seems time to pull in normalize.css
through other means. This change uses Rails Assets to get normalize.css.
It wraps Bower, which pulls directly through GitHub tags, so it should
always up-to-date.

Ref: markmcconachie/normalize-rails#23