Ignore GitHub status update errors
Sometimes the user's or Hound's token isn't privileged to set the
GitHub PR status for a given repo. Status isn't critical, and shouldn't
fail the rest of the build -- as long as we can access the files and
make comments.

We're actually seeing a lot of status "Not Found" errors that prevent
builds from finishing, thus this is measure to mitigate that.
This will reduce the job failure noise and allow us to identify which
repos we can't _actually_ review.