Fix intermittent test failure on config dropdown
There is an issue when a org-wide config is being enabled on the page.
When AJAX request completes we show a green checkmark and we use that in
the test to wait for AJAX to finish.

The issue shows up when the toggle is first enabled, which causes the
checkmark to be show (thus it's present on the page). Then when we
elect a repo from the dropwdown, which will remove the checkmark and
show it again when the request is done. Thus, we need to add step that
verifies/waits that the checkmark is gone when we select the config
repo.