Switch to chrome webdriver for JS feature tests
The future of QT and capybara-webkit is unknown.
Selenium with Chrome (headless) seems like a great alternative.

Ensure the repos list landing page (after sign in) finishes loading the
list of repose (via Ajax) for JS tests. Otherwise, a JS alert popup with
a failure causes tests to fail. This actually happens in Chrome on
production when navigating from the repos page before the AJAX request
 to get the repos finishes (the `Your repos failed to load.` alert).