Remove repo memberships that can't set the status
Sometimes users get removed from the repos, and they can no longer set
the status (or access repo files), so we should remove them from the
repo on our side.

There is one thing worth considering, if the repo is public but the user
isn't part of that repo, then the user can still access the files
(everything needed to review a PR and make a comment), but they can't
set the status. Should we remove them in that case? On one side, we
can't set the status but continue reviewing -- if no violations are
found, it'll look like Hound never did anything since the status won't
have anything from Hound. On the other hand, if we remove that user,
then we'll have to use Hound's token, thus incurring more API hits.