Ensure presence of current org-wide config option
If the current org-wide config repository is not in the list of the
repos (for whatever reason), it won't be included in the dropdown list
of possible repos for org-wide configuration. This will cause us to
select another/different repo for org-wide config automatically,
effectively changing the org-wide config on the user.

This ensures the current org-wide config is always in that list.