Remove user from repo if they can't set pr status
This will allow us to cleanup users who no longer have access to set the
status on a repo. Since we retry a failed job, we should get a good
token eventually.

There is a bit of a drawback, that we'll end-up using Hound's token more
for public repos, but that seems like a fair trade off.