Suppress error about failure to set GH status
This will reduce the noise of exceptions that we get.
We currently report a general Sentry notification when we can't set the
status.

Additionally, some open source repos cannot be reviewed if we don't have
a valid token for them (all user tokens have become invalid or expired),
because we can't set their status.
With this change, we'll be able to review them and make comments on
violations, even though we cannot set the status.