Don't raise if we can't set the status
Raising when we can't set the status will abort the review process, and
we have situations where we cannot set the status of a PR
(public repos in particular).

If we can't set the status, at least we can still make comments on
violations that are present in the code.